Rename pages directory to sections
To avoid conflict with Astro naming conventions
