+++ src/scene/Scene.tsx
@@ -1,7 +1,7 @@
 import { Environment, Preload } from "@react-three/drei";
 import { Suspense } from "react";
 import HDRI from "@assets/empty_warehouse_01_1k.hdr?url";
-import { pages } from "@pages/Pages";
+import { pages } from "@/sections/Pages";
 import { useLoadingState } from "@hooks/useLoadingState";
 import { CameraRig } from "./CameraRig";
 import { LoadingIndicator } from "./LoadingIndicator";
