+++ src/sections/android/Device.tsx
