+++ src/sections/android/icons/tap_band_icon.png
