+++ src/sections/intro/AvatarModel.tsx
