+++ src/sections/react/tictactoe/TicTacToePage.tsx
@@ -13,12 +13,12 @@ import {
   ReactTag,
   ThreeJsTag,
   TypescriptTag,
-} from "@pages/components/ProjectDescription";
+} from "@/sections/components/ProjectDescription";
 import { useHtmlPortal } from "@hooks/useHtmlPortal";
 import { useScreenState } from "@hooks/useScreenState";
 import { useSpringScaleVisibility } from "@hooks/useSpringScaleVisibility";
-import { PageComponentProps } from "@pages/Pages";
-import { useScrollPages } from "@pages/useScrollPages";
+import { PageComponentProps } from "@/sections/Pages";
+import { useScrollPages } from "@/sections/useScrollPages";
 import Font from "./Comfortaa_Bold.json";
 
 const XPiece = { color: "red", shape: "x" } as const;
