+++ vite.config.ts
@@ -12,7 +12,7 @@ export default defineConfig({
       "@assets": path.resolve(__dirname, "src/assets"),
       "@hooks": path.resolve(__dirname, "src/hooks"),
       "@components": path.resolve(__dirname, "src/components"),
-      "@pages": path.resolve(__dirname, "src/pages"),
+      "@sections": path.resolve(__dirname, "src/sections"),
       "@scene": path.resolve(__dirname, "src/scene"),
       "@analytics": path.resolve(__dirname, "src/analytics"),
     },
